Allow variants to be passed into all elements in P and S families
