+++ cpp/basix/e-brezzi-douglas-marini.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "cell.h"
+#include "element-families.h"
 #include "finite-element.h"
 
 namespace basix::element
@@ -12,9 +13,13 @@ namespace basix::element
 /// Create BDM element
 /// @param[in] celltype The cell type
 /// @param[in] degree The degree of the element
+/// @param[in] lvariant The Lagrange variant to use when defining the element to
+/// take integral moments against
 /// @param[in] discontinuous Controls whether the element is continuous or
 /// discontinuous
 /// @return A finite element
-FiniteElement create_bdm(cell::type celltype, int degree, bool discontinuous);
+FiniteElement create_bdm(cell::type celltype, int degree,
+                         element::lagrange_variant lvariant,
+                         bool discontinuous);
 
 } // namespace basix::element
