+++ cpp/basix/e-nedelec.cpp
@@ -154,6 +154,7 @@ xt::xtensor<double, 2> create_nedelec_3d_space(int degree)
 
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_nedelec(cell::type celltype, int degree,
+                                             element::lagrange_variant lvariant,
                                              bool discontinuous)
 {
   if (degree < 1)
@@ -188,18 +189,16 @@ FiniteElement basix::element::create_nedelec(cell::type celltype, int degree,
   }
 
   // Integral representation for the boundary (edge) dofs
-  FiniteElement edge_space
-      = element::create_lagrange(cell::type::interval, degree - 1,
-                                 element::lagrange_variant::legendre, true);
+  FiniteElement edge_space = element::create_lagrange(
+      cell::type::interval, degree - 1, lvariant, true);
   std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
       edge_space, celltype, tdim, 2 * degree - 1);
 
   // Face dofs
   if (degree > 1)
   {
-    FiniteElement face_space
-        = element::create_lagrange(cell::type::triangle, degree - 2,
-                                   element::lagrange_variant::legendre, true);
+    FiniteElement face_space = element::create_lagrange(
+        cell::type::triangle, degree - 2, lvariant, true);
     std::tie(x[2], M[2]) = moments::make_integral_moments(face_space, celltype,
                                                           tdim, 2 * degree - 2);
   }
@@ -219,7 +218,7 @@ FiniteElement basix::element::create_nedelec(cell::type celltype, int degree,
     {
       std::tie(x[3], M[3]) = moments::make_integral_moments(
           element::create_lagrange(cell::type::tetrahedron, degree - 3,
-                                   element::lagrange_variant::legendre, true),
+                                   lvariant, true),
           cell::type::tetrahedron, 3, 2 * degree - 3);
     }
     else
@@ -240,11 +239,13 @@ FiniteElement basix::element::create_nedelec(cell::type celltype, int degree,
 
   return FiniteElement(element::family::N1E, celltype, degree, {tdim}, wcoeffs,
                        x, M, maps::type::covariantPiola, discontinuous,
-                       degree - 1);
+                       degree - 1, {}, lvariant);
 }
 //-----------------------------------------------------------------------------
-FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
-                                              bool discontinuous)
+FiniteElement
+basix::element::create_nedelec2(cell::type celltype, int degree,
+                                element::lagrange_variant lvariant,
+                                bool discontinuous)
 {
   if (celltype != cell::type::triangle and celltype != cell::type::tetrahedron)
     throw std::runtime_error("Invalid celltype in Nedelec");
@@ -264,8 +265,8 @@ FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
       xt::xtensor<double, 3>({0, tdim, 0}));
 
   // Integral representation for the edge dofs
-  FiniteElement edge_space = element::create_lagrange(
-      cell::type::interval, degree, element::lagrange_variant::legendre, true);
+  FiniteElement edge_space
+      = element::create_lagrange(cell::type::interval, degree, lvariant, true);
   std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
       edge_space, celltype, tdim, 2 * degree);
 
@@ -273,7 +274,7 @@ FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
   {
     // Integral moments on faces
     FiniteElement face_space
-        = element::create_rt(cell::type::triangle, degree - 1, true);
+        = element::create_rt(cell::type::triangle, degree - 1, lvariant, true);
     std::tie(x[2], M[2]) = moments::make_dot_integral_moments(
         face_space, celltype, tdim, 2 * degree - 1);
   }
@@ -291,7 +292,8 @@ FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
     {
       // Interior integral moment
       std::tie(x[3], M[3]) = moments::make_dot_integral_moments(
-          element::create_rt(cell::type::tetrahedron, degree - 2, true),
+          element::create_rt(cell::type::tetrahedron, degree - 2, lvariant,
+                             true),
           celltype, tdim, 2 * degree - 2);
     }
     else
@@ -314,6 +316,7 @@ FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
   }
 
   return FiniteElement(element::family::N2E, celltype, degree, {tdim}, wcoeffs,
-                       x, M, maps::type::covariantPiola, discontinuous, degree);
+                       x, M, maps::type::covariantPiola, discontinuous, degree,
+                       {}, lvariant);
 }
 //-----------------------------------------------------------------------------
