+++ test/test_nedelec_second_kind.py
@@ -142,14 +142,14 @@ def sympy_nedelec(celltype, n):
     return g
 
 
-# TODO: turn this test back on when a equispaced Nedelec variant is available
-@pytest.mark.parametrize("order", [1, 2, 3])
-def xtest_tri(order):
+@pytest.mark.parametrize("degree", [1, 2, 3])
+def test_tri(degree):
     celltype = basix.CellType.triangle
-    g = sympy_nedelec(celltype, order)
+    g = sympy_nedelec(celltype, degree)
     x = sympy.Symbol("x")
     y = sympy.Symbol("y")
-    nedelec = basix.create_element(basix.ElementFamily.N2E, basix.CellType.triangle, order)
+    nedelec = basix.create_element(
+        basix.ElementFamily.N2E, basix.CellType.triangle, degree, basix.LagrangeVariant.equispaced)
     pts = basix.create_lattice(celltype, 6, basix.LatticeType.equispaced, True)
     nderiv = 3
     wtab = nedelec.tabulate(nderiv, pts)
@@ -166,15 +166,15 @@ def xtest_tri(order):
             assert(numpy.isclose(wtab[basix.index(kx, ky)], wsym).all())
 
 
-# TODO: turn this test back on when a equispaced Nedelec variant is available
-@pytest.mark.parametrize("order", [1, 2])
-def xtest_tet(order):
+@pytest.mark.parametrize("degree", [1, 2])
+def test_tet(degree):
     celltype = basix.CellType.tetrahedron
-    g = sympy_nedelec(celltype, order)
+    g = sympy_nedelec(celltype, degree)
     x = sympy.Symbol("x")
     y = sympy.Symbol("y")
     z = sympy.Symbol("z")
-    nedelec = basix.create_element(basix.ElementFamily.N2E, basix.CellType.tetrahedron, order)
+    nedelec = basix.create_element(
+        basix.ElementFamily.N2E, basix.CellType.tetrahedron, degree, basix.LagrangeVariant.equispaced)
     pts = basix.create_lattice(celltype, 6, basix.LatticeType.equispaced, True)
     nderiv = 1
     wtab = nedelec.tabulate(nderiv, pts)
